#ALSA-12 test promises with callbacks
Promise is created by invoking the `Promise` c-tor function with `new`.
We need to call the `resolve` method when promise is fulfilled,
and the `reject` method when it is rejected.
After the promise is obtained, we will use it by supplying
success and failure callbacks to it's `then` method.

Promises can be chained if the success/failure callbacks return promises